Refactor Entities to have pointer to its EntityManager.

+++ WeSp3D/Engine/SceneManager/EntityManager/ComponentManager/Component/Model/SkeletalModel.cpp
@@ -534,7 +534,7 @@ void SkeletalModel::boneTransform(float timeInSeconds, std::vector<glm::mat4>& T
   float TicksPerSecond = (float)(m_pScene->mAnimations[0]->mTicksPerSecond != 0 ? m_pScene->mAnimations[0]->mTicksPerSecond : 25.0f);
 
   float TimeInTicks = timeInSeconds * TicksPerSecond;
-  float AnimationTime = fmod(TimeInTicks, animDuration);
+  float AnimationTime = static_cast<float>(fmod(TimeInTicks, animDuration));
 
   ReadNodeHeirarchy(AnimationTime, m_pScene->mRootNode, Identity);
